+++ DowndetectorMCP.API/DowndetectorMCP.API.csproj
@@ -10,4 +10,8 @@
     <PackageReference Include="Microsoft.Playwright" Version="1.57.0" />
   </ItemGroup>
 
+  <ItemGroup>
+    <ProjectReference Include="..\DowndetectorMCP.Shared\DowndetectorMCP.Shared.csproj" />
+  </ItemGroup>
+
 </Project>
